move Column.java back to impl directory, remove orphan code
